% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huc12_full.R
\docType{data}
\name{huc12_full}
\alias{huc12_full}
\title{Outline of the Mississippi Sound Watershed HUC12s}
\format{
A Simple feature collection with 567 features and 6 fields.
\describe{
\item{\code{huc12}}{character, HUC ID.}
\item{\code{name}}{character, name of the HUC unit.}
\item{\code{hutype}}{character, type of the HUC unit.}
\item{\code{hutype_description}}{character, description of the HUC unit type (i.e. Frontal, Island, Standard, Water).}
\item{\code{humod}}{character}
\item{\code{tohuc}}{character, which HUC12 the water from this one goes into next.}
\item{\code{geometry}}{POLYGON of the MSEP watershed HUC12s.}
}
}
\source{
National Hydrography Dataset, as downloaded from USGS on 6/4/2025 from \url{https://prd-tnm.s3.amazonaws.com/StagedProducts/Hydrography/NHD/State/GPKG/NHD_H_Mississippi_State_GPKG.zip}.
}
\usage{
huc12_full
}
\description{
The Mississippi Sound Watershed is made up of HUC4s 0317 (Pascagoula River basin) and 0318 (Pearl River basin). This object contains the outlines of the HUC12s within both basins.
}
\details{
The raw data were subsetted to the geopackage file \code{msep_hucs.gpkg} using the script \code{subset_NHD.R}. The subset was further processed into this final form, using the \code{tigris} package for state and cartographic boundary delineations, in \code{make_boundaries.R}.
Both scripts and the subsetted geopackage file are available in the \code{data-raw} directory of the package repository at:
\url{https://github.com/CMEP-MS/msepBoundaries}.
}
\keyword{datasets}
