% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subbasins_full.R
\docType{data}
\name{subbasins_full}
\alias{subbasins_full}
\title{Subbasins (HUC8s) of the full Mississippi Sound Watershed}
\format{
A Simple feature collection with 14 features (rows) and 2 fields.
\describe{
\item{\code{huc8}}{character, HUC8 ID.}
\item{\code{name}}{character, name of the subbasin.}
\item{\code{geometry}}{POLYGON of each subbasin.}
}
}
\source{
National Hydrography Dataset, as downloaded from USGS on 6/4/2025 from \url{https://prd-tnm.s3.amazonaws.com/StagedProducts/Hydrography/NHD/State/GPKG/NHD_H_Mississippi_State_GPKG.zip}.
}
\usage{
subbasins_full
}
\description{
The Mississippi Sound Watershed is made up of HUC4s 0317 (Pascagoula River basin) and 0318 (Pearl River basin). This is the combination of all HUC8 units within them.
}
\details{
The raw data were subsetted to the geopackage file \code{msep_hucs.gpkg} using the script \code{subset_NHD.R}. The subset was further processed into this final form, using the \code{tigris} package for state and cartographic boundary delineations, in \code{make_boundaries.R}.
Both scripts and the subsetted geopackage file are available in the \code{data-raw} directory of the package repository at:
\url{https://github.com/CMEP-MS/msepBoundaries}.
}
\keyword{datasets}
