% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\docType{package}
\name{msepBoundaries-package}
\alias{msepBoundaries}
\alias{msepBoundaries-package}
\title{msepBoundaries: Provides Spatial Boundaries for the MSEP Area}
\description{
Provides several versions of boundary files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CMEP-MS/msepBoundaries}
  \item Report bugs at \url{https://github.com/CMEP-MS/msepBoundaries/issues}
}

}
\author{
Kim Cressman
}
\keyword{internal}
